Unsplash API Implementation
Note: Might need to temporarily hard code the access ID & secret for Unsplash API to work (it's currently using dotenv module to hide)

Implemented Unsplash API to add random image icon for user logging in.

Checks if the user has an icon, and pings Unsplash for a random image if not.

Made changes to DB to add "userIcon" to dictionary.

Added image to reminder/index.ejs & created icon class in public/stylee.css